@@ -1,12 +1,16 @@
-module.exports = {
-  presets: [
-    [
-      '@babel/preset-env',
-      {
-        useBuiltIns: 'usage',
-        corejs: 3
-      }
-    ]
-  ],
-  plugins: ['@babel/plugin-transform-runtime', '@babel/plugin-transform-modules-commonjs']
-}
+const presets = [
+  [
+    '@babel/env',
+    {
+      targets: {
+        edge: '17',
+        firefox: '60',
+        chrome: '67',
+        safari: '11.1'
+      },
+      useBuiltIns: 'usage'
+    }
+  ]
+]
+
+module.exports = { presets }
