@@ -1,3 +1,9 @@
 module.exports = {
-  ddd: 'aaa'
+  first: 'bai',
+  last: 'bai',
+  id: 66,
+  ids: ['5632953c4e345e145fdf2df8', '563295464e345e145fdf2df9'],
+  add: (a, b) => {
+    return a + b
+  }
 }
